requirements.txt
typer>=0.9.0
rich>=13.3.5
pyyaml>=6.0
python-dotenv>=1.0.0
openai>=1.3.0
requests>=2.28.2
gitpython>=3.1.31
aiohttp>=3.8.0
lightning>=2.0.0
